+++ packages/chrome-devtools/src/utils/chrome/storage.ts
@@ -6,21 +6,14 @@ import {
 } from '../sdk';
 import { injectScript } from './index';
 
-export const mergeStorage = async (...args: any[]) => {
-  await injectScript(mergeLocalStorage, false, ...args);
-};
+export const mergeStorage = async (...args: any[]) =>
+  injectScript(mergeLocalStorage, false, ...args);
 
-export const removeStorageKey = async (
-  tab: chrome.tabs.Tab,
-  ...args: any[]
-) => {
-  await injectScript(removeLocalStorageKey, false, ...args);
-};
+export const removeStorageKey = async (...args: any[]) =>
+  injectScript(removeLocalStorageKey, false, ...args);
 
-export const removeStorage = async (tab: chrome.tabs.Tab, ...args: any[]) => {
-  await injectScript(removeLocalStorage, false, ...args);
-};
+export const removeStorage = async (...args: any[]) =>
+  injectScript(removeLocalStorage, false, ...args);
 
-export const setStorage = async (tab: chrome.tabs.Tab, ...args: any[]) => {
-  await injectScript(setLocalStorage, false, ...args);
-};
+export const setStorage = async (...args: any[]) =>
+  injectScript(setLocalStorage, false, ...args);