+++ packages/chrome-devtools/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @module-federation/devtools
 
+## 0.0.26
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.10
+
 ## 0.0.25
 
 ### Patch Changes